requirements.txt
###### Requirements without Version Specifiers ######
flask
pyjwt
gunicorn
flask-cors